entities and service

+++ src/main/java/com/carlosc/crud/example/Product.java
@@ -9,12 +9,12 @@ import javax.persistence.Id;
 public class Product {
     private Integer id;
     private String name;
-    private float price;
+    private Double price;
 
     public Product() {
     }
 
-    public Product(Integer id, String name, float price) {
+    public Product(Integer id, String name, Double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -38,11 +38,11 @@ public class Product {
         this.name = name;
     }
 
-    public float getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
 }
